Change EncryptCookies::serialized behaviour
Currently, the serialize method returns a bool value no matter what is in the `$name` parameter. I changed it to work as `in_array` check just like the `isDisabled' method for the encryption. Now you can add the cookie's names to the `$serialize` static property which you want to serialize.