Fix replacing required :input with null on PHP 8.1
Fixes an issue where data lacking one or more of the fields under
validation will cause tests to fail with an `ErrorException`.

That exception is a PHP deprecation warning triggered by the call to
`str_replace()` inside `replaceInputPlaceholder()`, which assumes the
value returned from `getDisplayableValue()` will always be a string.