Remove unnecessary check if clone->columns exists as it's just for pagination.
This was added on this PR https://github.com/laravel/framework/pull/32691 at this specific commit:
https://github.com/laravel/framework/commit/98a242e21041462054b965e587c250ac7be4f912

A part from that, there is no way to select null columns, always
receiving array(null) instead.

On the other hand, if we join tables and use select *, we could be using
diferent fields from each table but it always breaks when paginating because of
duplicate column name 'id'.