Allow transaction retries when savepoints disabled
In my use case, I disabled MySQL savepoints to allow DB::transaction() to retry in deadlock scenarios. However, the current logic throws the exception even though it's seemingly not necessary when savepoints are disabled. Since there is only one transaction open in MySQL, the entire transaction can be rolled back and replayed. So we check to see if the query grammar supports savepoints before throwing.