implement `has` method for Cache stores
Currently the `Repository` is responsible for determining if a key exists in the cache. It does this be trying to get the keyed value from the store, and checking if it is `null`. We'll switch to allow each store to define its own `has` method to determine if a key exists.

I've done the easy store implementations here. The others are a little more difficult, and will have to dig more into them.