Add a new Collection#containsAll method
This commit adds a new containsAll
method to both Collection and
LazyCollection, and some tests
to make sure it works correctly.

This method allows to determine if
a smaller array/collection items are
all present in a laravel collection.

* Why is that needed?

I think this is clearly something Laravel
collections lacks of.
In my case, I need to check if a laravel
collection contains all the stuff a user
sent to the controller.
I've got something like an enterprise
admin panel (not to be confused with
the global admin panel) which allows the
administrator to manage the employees
permissions.
Since the enterprise administrator is
sending ids via a form, I must verify
all the ids provided are all their own
employees, to avoid malicious users to
revoke access to employees of other
enterprises by sending random ids for
example.

I think this use case is not a niche one,
but a relatively current one.

In order to illustrate my use case, here
is a snippet of code based on it:
```PHP
if (!$this->containsAll($enterpriseEmployeesIds, $employeesId)) {
    $flash->error("Specified employees does not belong to your enterprise!");
    return back();
}
```
with $flash being an instance of an utility
class wrapping
`$request->flash('whatever', [logLevel, msg])`.

* Implementation notes

At first, I plan to use LazyCollection#passthru
for the LazyCollection#containsAll implementation
but I wasn't able to make it work.
Then, I realize that copying the code of the
Collection#containsAll implementation is the thing
to do.
The code does not behave the same with a
LazyCollection as it does with a Collection
in terms of memory usage.
Basically, the difference is that calling passthru
creates a Collection, which is entirely allocated
in memory.
Furthermore, the creation of an intermediary
Collection adds unnecessary overhead and would forbid
the usage of containsAll with large LazyCollections.

I'm not familiar with the laravel internals, so the
implementation is pretty simple to understand without
specific knowledge about these.

I tried to test it the best I can, but if you want me to
add more tests for some weird edge cases for example,
don't hesitate to ask.

* Usage

```PHP
$collection->containsAll($arr);
```
with $arr being an array, a Collection, a LazyCollection
or whatever implements Enumerable.

It will crash if $arr is null because we can't do a
foreach on a null element.
I think it is the intended behavior and that it could
just be stated in the documentation.