Add a new Collection#containsAll method
This commit adds a new containsAll
method to both Collection and
LazyCollection, and some tests to
make sure it works correctly.

This method allows to determine if
a smaller array/collection items are
all present in a laravel collection.

* Usage

```PHP
$collection->containsAll($arr);
```
with $arr being an array, a Collection, a LazyCollection
or whatever works with getArrayableItems.

When $arr is empty or null, containsAll returns true.

* Why is that needed?

I think this is clearly something Laravel
collections lacks of.
In my case, I need to check if a laravel
collection contains all the stuff a user
sent to the controller.
I've got something like an enterprise
admin panel (not to be confused with
the global admin panel) which allows the
administrator to manage the employees
permissions.
Since the enterprise administrator is
sending ids via a form, I must verify
all the ids provided are all their own
employees, to avoid malicious users to
revoke access to employees of other
enterprises by sending random ids for
example.

I think this use case is not a niche one,
but a relatively current one.

In order to illustrate my use case, here
is a snippet of code based on it:
```PHP
if (!$this->containsAll($enterpriseEmployeesIds, $employeesId)) {
    $flash->error("Specified employees does not belong to your enterprise!");
    return back();
}
```
with $flash being an instance of an utility
class wrapping
`$request->flash('whatever', [logLevel, msg])`.

* Implementation notes

The implementation of containsAll is
based on the one suggested by
@JosephSilber in my precedent pull
request (#36915).

For some reason, I didn't manage to
make it work with array_splice, but
it works with unset, so I used unset.

The code behaves the same with a
LazyCollection as it does with a Collection
because the containsAll implementation is
the same, and located in EnumeratesValues
trait.

I tried to test it the best I can, but if you want me to
add more tests for some weird edge cases for example,
don't hesitate to ask.