Change order of 'deleting' event
It is my humble opinion that. The 'deleting' event should be fired before the item gets deleted from the database. Not after
Before this PR pivot::deleting would not have access to the Primary key  because its parent class usually model::delete already deleted so any listener will not have access to the full pivot table entry. because it would not exist in the database.