Fix fields not required with required_unless
This fixes an issue where the `required_unless` rule does not add a
"required" error when the 'other' field isn't included in input data.

Take `'field' => 'required_unless:other,true'` as an example. With `[]`,
the absense of `other` implies its value is *not `true`* and therefore
`field` should be required. Instead, validation automatically passes.

Flipping `return true` fixed the failing test but failed with valid data
that *wasn't* tested. The extra two tests ensure it still passes if the
field is required *and provided*, or absent when 'other' *expects* null.