Create getter for the middleware priority
This getter will help to create tests to ensure that middlewares are
registered in the right priority.

My use case for this is a middleware that I want to ensure gets
registered after the `StartSession` but right before the
`AuthenticatesRequests` one. With this getter I can use their indexes to
write a test for it.