Update return type for tap
`when` already has `$this|mixed` so this should be reflected in `tap` on `BuildsQueries`. If you look at the logic inside `when` you'll see that this is the correct return type as it's either the return value of the given callback or `$this`. A previous attempt was made here: https://github.com/laravel/framework/pull/32717

See a more thorough explanation here: https://github.com/laravel/framework/issues/38343