Add `hasTables` function
Sometimes we need to load the data in `AppServiceProvider` before the view has been rendered as the following code shows.

```PHP
$users = \App\Models\User::get();

\Illuminate\Support\Facades\View::share('users', $users);
```
So, what if the developer NOT migrates the database yet? An error will be triggered because the `users` table does NOT exist yet, Here we want to check if the table exists or NOT using `hasTable` function. So, what if we want to load data from many tables in the `AppServiceProvider` we will need to check the existence of all of them as the following code shows

```PHP
if(Schema::hasTable('users') && Schema::hasTable('cats') && ...) {
   // get all data from these tables here
}
```

The form of code here is NOT formatted well. So, `hasTables` function will be here for that reason.