[8.x] Add bad request test response assertion
Inspired by PR #38553.  A readability enhancement in the same manner.
Just like `assertOk`, `assertUnauthorized` and company this adds an
`assertBadRequest` to ensure that a [400 Bad Request](https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.1) was sent
in the response.

The intention is to increase readability and add broader support to the
existing suite of response code assertions which are much easier to
understand.

This introduces no breaking changes since it does not modify any
existing methods.

Also introduced a test case for the assert error path and the path where
there is no assertion error.