[8.x] Use Application::storagePath() instead of global function in artisan cache:clear
Since `illuminate/cache` is distributed as a separate package, it should probably not rely on the existence of global function `storage_path()`. Instead, it should rely on `Application::storagePath()` since that method is part of the `Application` contract.

This is consistent with the use of [`Application::databasePath()` in `CacheTableCommand`](https://github.com/laravel/framework/blob/8.x/src/Illuminate/Cache/Console/CacheTableCommand.php#L77)

Co-Authored-By: QWp6t <QWp6t@users.noreply.github.com>