Inform the application container about deferred providers before registering providers
This change informs the application container about the deferred providers before registering the eager loaded providers. This fixes the issue of the vague exception of "Facade root not set" when attempting to use a deferred service provider via a facade (such as Cache) during the registration process of eagerly loaded service providers.