Support adding column "after" other new column
Previously, new columns could only be added after existing columns, so
referencing a new column with the "after" modifier caused a MySQL error.

From now one, referencing a new column causes the column definition to
be inserted simply after the definition of the referenced column.

In particular, this allows us to create a new table with `timestamps()`,
and adding a `created_by` column after `created_at` at the same time.