Update SchemaState Process to remove timeout
As shown in Symphony Process docs: https://symfony.com/doc/current/components/process.html#process-timeout

All Process spawned have a default timeout of 60 seconds. For a pg_dump script. 60 seconds might or might not be enough especially if this is run in a CI/CD envrionment where process contention (in shared or virtual hardware) is an issue.

Removing the timeout is great because running migrations could be seen as a long lived process by most users and developers.