chore (logging): support action_level configuration
Enable the usage of the action_level configuration for log channels. When the action_level configuration is used the choosen handler by the developer will be wrapped in a Fingerscrossed handler.

The Fingerscrossed handler will buffer all log records that are at or above the provided level configuration. Only when a log record is provided at or above the action_level all the buffered records will be flushed to the log itself.

This methodology helps developers to gather debug information about a process while only logging that debug information if the code hits for example a critical error.

@see https://github.com/Seldaek/monolog/blob/main/src/Monolog/Handler/FingersCrossedHandler.php
@see https://symfony.com/doc/current/logging.html#handlers-that-modify-log-entries