Resolve  Blueprint out of the container
Currently, it is almost impossible to swap the `Blueprint $table` implementation for migrations with a fake one from within a "package space".
Because the `Schema` facade returns a new object each time from the `getFacadeAccessor` method (hence there is no way of mocking or hot swapping) and non of the other objects are resolved out of the container. (Connection objects or Builder objects)
There is also no change of using the `resolver` property on the `Builder` class from a package space because if you set it before the migration starts it will be thrown away after the `Schema` is called in the migration.

Currently "illuminate/database" depends on "illuminate/container" so we can use the Container class directly.
This way it is possible to swap the Blueprint class upon the `registration phase`.