[Database] Allow the creation of char columns with length 0
The most effective way, storage-wise, to store boolean values in the database, in MySQL, is to use a `char(0)` column that will store NULL for false and "" (empty string) for true. So allow the creation of a char column with the length of zero.

Note: This is only useful, and true, for MySQL. For PostgreSQL, for example, it's not.