Don't add duplicate callbacks
While troubleshooting a semi-related problem with tests, Livewire, and the `TrimStrings` middleware, I noticed that while running a batch of tests, callbacks will "stack up"....that is, they will keep being added.

I'm not sure that this behavior is necessarily bad, and it may actually be a good thing under some circumstances...I'm not sure.  But, it seemed to me that it would generally be considered something worth avoiding.

To test this, I simply dumped the count of `static::$skipCallbacks` at the end of `skipWhen()` and, when running my tests normally, the value just continues to increase.

So...I put this little change in anticipating that it might make tests a little more efficient.  🤓 

Please let me know if there are any questions/concerns...or let me know if I've missed something.  😜 

Thx.  🚀