Rollback a specific batch of migrations - Add a new option `batch` to the RollbackCommand. - Add a new function `getMigrationsByBatch` to the MigrationRepositoryInterface. - Implement `getMigrationsByBatch` in DatabaseMigrationRepository.
This new option allows the developer to rollback a specific batch of migrations. This can be helpful
in situations where a production bugfix introduced database changes after the current feature branch introduced database changes. The develop may not be able to rollback the production bugfix batch, but still wants to revert the batch from the current feature branch to make slight changes to the database migrations introduced in the current feature branch.

A PR to laravel/docs will also be submitted.