Make throttle lock acquisition attempt retry time configurable
Depending on context, 750ms can be a long time to wait between retries
for a throttle.  We are making the retry time configurable so that users
can tune the throttling further.

All changes are backwards compatible.

I set the unit as milliseconds because this seemed like the natural time
unit to use.  Seconds would require floats which felt dirty.  That being
said, it's not consistent with the unit used for `->every` and `->block`
on the builder.  So I'm open to changing it for consistency.

```php
Redis::throttle('my-key')
    ->allow(5)
    ->every(10)
    ->retryWait(100)
    ->then(static fn () => echo "hello!");
```