Bugfix #41302 - Alternative
This is an alternative option to PR #41502, and adds a match
for 'socket' ("socket error when (reading|writing)") to trigger
a reconnection.

I am not sure if this is valid in other languages, too.

My personal preference would be to remove the entire if and
ALWAYS reconnect on any Exception

Signed-Off-By: Rob Thomas xrobau@gmail.com