Make reconnectIfMissingConnection public
I've use-cases where I need to get the PDO from a connection but also
need to respect, in case a disconnect happened, to perform a reconnect.

Currently, I've to:
- get pdo
- check if null
- call reconnect

everywhere I need it. Or I could just call this one on the connection
object, if it were public, and then get the PDO.