Update CompilesErrors.php
This update allows for the @error blade directive to accept an array of keys to check for in the MessageBag.

When given an array, it will check if the error bag has an error in any of the keys, if so, the first error in the first key that has errors will be available in the `$message` variable.

Use case:
```php
<input id="accessCode" type="text" class="form-control @error(['code', 'throttled']) is-invalid @enderror"
placeholder="Code" name="code" value="{{ old('code') }}" aria-describedby="accessCodeFeedback">
<label>{{ trans('labels.access_code') }}</label>
@error(['code', 'throttled'])
<div id="accessCodeFeedback" class="invalid-feedback">{{ $message }}</div>
@enderror
```

I'm not sure on how to add testing coverage to this. Some help/guidance is welcome.