[9.x] - Cast locale to avoid null
Context: #43125 

This is causing the warning, at least, maybe there are more places. I know it's needed to pass a string here, so as a first action, i added a strval to the calls, which didn't helped. As this is solved with casting i wouldn't expect issues and hope you people are fine with it.