fix: Fix empty arrays on request bodies
This PR fixes a problem I've encountered some times. The problem is that on some APIs, an empty array as the request's body may lead to false errors regarding parsing the payload. While it may seem that sending nothing on an Http::post($url) method may suffice it, it still sends an empty "[]" array.