create `EmptyArgument` class
`null` can have different meanings when used in different contexts. in a lot of places in the framework, it is used to mean "ignore this parameter when no argument is passed, and execute different behavior than if a value had been passed". for the most part this is fine, but it limits us slightly by not be able to check for a desired `null` value. this new empty marker class allows us to specifically designate we wish to treat the method call as if the argument was not passed, without conflating the use of `null`.