do not trim the request URL
URLs with and without a trailing slash are technically different routes according to the browser, so we should not be auto trimming the trailing slash here.

the `url()` function also does its own trimming, so this commit doesn't actually solve the problem, but it does get rid of the duplicate trimming. changing `url()` has wider reaching effects, so will have to look closer into that change.