chore (logging): add stop_buffering config option (#44071)
Adding the stop_buffering config option.

The FingersCrossedHandler Logger will first start buffering log records until a record exceeds the configured action_level. When the action_level is exceeded the buffered records will be flushed to the provided log location.

By default the FingersCrossedHandler will stop buffering log records after its first flush (due to the $stopBuffering property). All records pushed to the log handler afterwards will be flushed directly to the log location.

With this config update the log handler can be configured to resume the buffering after a flush occurs.

This can be useful for applications that rely heavily on gather debug log records which only are useful when an error logging is passed. In the old setup all the debug records after an error record would be flushed to the log location.

@see https://github.com/Seldaek/monolog/blob/main/src/Monolog/Handler/FingersCrossedHandler.php#L73

Co-authored-by: Wim Ulkeman <wulkeman@assuradeurengilde.nl>