Solved the problem of incorrect parameter definition.
By default, the parameter should listen to the route parameter, not the model name. 
Documentation (Authorizing Resource Controllers)[https://laravel.com/docs/9.x/authorization#authorizing-resource-controllers] recommends specifying the parameter initially. The second parameter is optional. Ларавел will figure it out for you. In 99% of cases it works. I was lucky to be in the 1%.
Below is a complete description of the problem.