Adds getToken method to Request instances
Getting the token from the request (query, body, and headers) is universal and may very well be needed elsewhere (not just the TokenGuard instance). It would be good if this was available (and didn't depend on the 'token' guard being set) but if not, developers could just add as a macro.