Add Dumpable contract
Many Laravel classes have `dump` and `dd` methods.

It would be nice if these methods would be called when an object of this
sort is passed to `dd` or `dump`, instead of the default behavior. This
would be especially useful when passing multiple values to the dump
methods.

Added a new contract, `Dumpable`, which requires the presence of the
`dump` and `dd` methods on the implementing class.

Since `dd` and `dump` are Symfony functions, copied over to
`helpers.php` and modified them. These should get loaded before the
Symfony ones.

Will follow up with tests.