Make `collect()` helper consistent with `new Collection()`
Even though using `null` and `[]` as argument produce the same result, syntactically they are different as PHPStan is unable to infer template types when `null` is passed. 

Currently when using `collect()` PHPStan will report these errors (as seen in larastan issue https://github.com/nunomaduro/larastan/issues/1115):
```
Unable to resolve the template type TKey in call to function collect
Unable to resolve the template type TValue in call to function collect
```

As a workaround you either have to explicitly pass empty array `collect([])` or use `new Collection()`.
I'm not sure if this is considered breaking BC or not.

I tried searching for existing issue/PR regarding this, but couldn't find anything. I apologize if this was addressed and declined already.