remove `HandlesAuthorization` trait from stub
the function in these traits are for the most part proxies to static methods on the `Response` class.

Even the [documentation](https://github.com/laravel/docs/commit/0682ca014e5e1768066ef93b4269a83d82978a70) doesn't use them, and instead goes directly to the `Response` class, so I think it would be best to remove them from the stubs.

I would say we should deprecate the trait as well, to encourage the documented use, but I'll hold on that.