Make createIndexName function public
Having this function public is useful to create a default index name according to the standard without creating your own function.  This is useful for checking if an index already exists.

See https://stackoverflow.com/a/45883428/6383358 for an example of a use case.

Also cleaned up function definition for later versions of Php.