allow serialization customization of relationships
currently you can set which attributes are "visible" or "hidden" when an Eloquent Model is serialized, and you can also set whether a relationship is visible or hidden.  but you cannot set the visibility of the individual attributes of your relationships at call time.  this change allows users to do that.

this can be activated by using dot notation

```php
$product->setVisible(['id', 'name', 'price', 'manufacturer', 'manufacturer.id', 'manufacturer.name']);
```