allow override of the Builder `paginate()` total
this allows the user to set the total number of results a query returns.  If the user provides this argument, the `paginate()` method will skip running its query to determine the total row count.

currently handled as a dynamic argument, since we cannot change the method signature on a patch or minor version.