fail when `assertSee` is passed an empty value
currently when passing an empty value to `assertSee` the test will pass.  these values include:

- `null`
- `[]`
- `''`
- `['foo', '', 'bar']`

rarely (IMO) is this the desired behavior, and may create false confidence in the test.

this change will fail the test by default when passed empty values, but also allow users to override if appropriate.