[10.x] Configurable storage path via environment variable
In the current implementation, the storage directory is statically set to `storage`. Given that this directory requires write permissions, it poses challenges in serverless environments and other scenarios where the filesystem might be read-only.

To enhance Laravel's adaptability to diverse hosting environments, this pull request introduces the capability to override the default storage directory using the `STORAGE_PATH` environment variable.