[10.x] Update MemcachedStore To Handle TTL Correctly
This PR was closed, but please note that the timestamp is treated as seconds which causes the TTL to not expire at the expected time. This causes huge issues when using memcached as the cache driver. Please check again

I fixed the issue by passing the expected arg (in seconds) to memcached