add function to get faked events
this function will return an array of all Events that have been faked.

this can be helpful when originally writing the tests to get a grasp of which events you may need to intercept.

both the `QueueFake` and `NotificationFake` have similar methods to let you see what has been intercepted by the fake.

https://github.com/laravel/framework/blob/11.x/src/Illuminate/Support/Testing/Fakes/QueueFake.php#L512
https://github.com/laravel/framework/blob/11.x/src/Illuminate/Support/Testing/Fakes/NotificationFake.php#L357