Events Dispatcher performance improvements
The event dispatching pipeline can be broken down into two implementation paths: Event as string and event as object.

Refactored dispatching to account for optimizations that can be made for string-based events and avoid running code when no listeners are registered.

For events that are registered, created a listeners cache to reduce repeated calls to makeListener() when the same event is called more than once.

Lastly, modified the shouldBroadcast calls to take the event object rather than the payload as that implementation seemed to be based on the presumed parsing of the object-based event calls.  Tests updated accordingly.